fix: execForStdOut should return complete output
Previously, the implementation register a listener to `exec` to get a buffer.
But I'm not sure if the buffer contains the complete output or it's just the
first buffer and there might be more.  There's a function in @action/exec:
`getExecOutput`. the new function returns a string and not a buffer.

Fix #47